Use 32-bit types instead of 16

+++ Primes/primes.c
@@ -14,13 +14,13 @@ typedef unsigned char bool;
 #endif
 
 #define NUM_PRIMES 4000
-uint16_t primes[NUM_PRIMES];
+uint32_t primes[NUM_PRIMES];
 
 
-bool primetest(uint16_t *primes, uint16_t *endprime, uint16_t candidate) {
-  for(uint16_t *ii=primes; ii<endprime; ++ii) {
-    uint16_t prime = *ii;
-    uint16_t quot = candidate / prime;
+bool primetest(uint32_t *primes, uint32_t *endprime, uint32_t candidate) {
+  for(uint32_t *ii=primes; ii<endprime; ++ii) {
+    uint32_t prime = *ii;
+    uint32_t quot = candidate / prime;
     if (quot < prime)
       return TRUE; 
     if (quot * prime == candidate)
@@ -30,17 +30,17 @@ bool primetest(uint16_t *primes, uint16_t *endprime, uint16_t candidate) {
 }
 
 
-void primecalc(uint16_t *primes, uint16_t num) {
+void primecalc(uint32_t *primes, uint32_t num) {
   if (num == 0) {
     return;
   }
 
   *primes = 2;
-  uint16_t *maxprime = primes + num;
-  uint16_t *current_prime = primes + 1;
-  uint16_t *primes_plus_1 = current_prime;
-  uint16_t candidate = 3;
-  uint16_t nprimes = 1;
+  uint32_t *maxprime = primes + num;
+  uint32_t *current_prime = primes + 1;
+  uint32_t *primes_plus_1 = current_prime;
+  uint32_t candidate = 3;
+  uint32_t nprimes = 1;
 
   while(current_prime < maxprime) {
     if (primetest(primes_plus_1, current_prime, candidate)) {
@@ -48,7 +48,7 @@ void primecalc(uint16_t *primes, uint16_t num) {
       ++current_prime;
       ++nprimes;
       if ((nprimes & 3) == 0) {
-        printf("%u     %u     %u     %u     %u\n",
+        printf("%lu     %lu     %lu     %lu     %u\n",
                *(current_prime - 4), *(current_prime - 3),
                *(current_prime - 2), *(current_prime - 1),
                getTimer());
